Add notify_immediate_close() to forcibly remove an fd from kqueue changelist.
The scenario is a library (libcurl) that is going to call close(2) immediately
before it removes from a queue.  On the kqueue backend that causes us to leave
a stale file descriptor with EV_DELETE in the pending changelist, to be cleared
on the next loop iteration.  That could potentially be reclaimed by an unrelated
fd, which is bad times (typical EBADF race).