StreamSocket: add Close() api.
One problem being tackled here is to flush any pending writes before closure.
We do this by having Write(nullptr, 0, ...) with a completion callback that
does the real close.

This may not work for the Windows named pipe hacks.  Need to look into that.

Another gap is that we don't have a clean TLS shutdown with notify-close.
To be addressed later.